requirements.txt

pandas
numpy
scikit-learn
streamlit
snowflake-connector-python
dbt-core
dbt-snowflake
sqlalchemy
pyyaml
shap
python-dotenv
plotly
